+++ app/src/main/java/io/github/muntashirakon/AppManager/history/ops/OpHistoryManager.java
@@ -41,9 +41,11 @@ public final class OpHistoryManager {
     public static final String HISTORY_TYPE_INSTALLER = "installer";
     public static final String HISTORY_TYPE_PROFILE = "profile";
     public static final String HISTORY_TYPE_CLEANUP = "cleanup";
+    public static final String HISTORY_TYPE_SINGLE_APP_ACTION = "single_app_action";
 
     @Retention(RetentionPolicy.SOURCE)
-    @StringDef({HISTORY_TYPE_BATCH_OPS, HISTORY_TYPE_INSTALLER, HISTORY_TYPE_PROFILE, HISTORY_TYPE_CLEANUP})
+    @StringDef({HISTORY_TYPE_BATCH_OPS, HISTORY_TYPE_INSTALLER, HISTORY_TYPE_PROFILE, HISTORY_TYPE_CLEANUP,
+            HISTORY_TYPE_SINGLE_APP_ACTION})
     public @interface HistoryType {
     }